refactor(team): /random を /team サブコマンドに統合
- /random コマンドを削除
- /team balance: ランクによる実力差を考慮したチーム分け
- /team random: 完全ランダムでチーム分け
- 共通ロジックを shared.ts に分離
